refactor(admin,synthesize): deriver client + verdict module + bounded RETRIEVAL_ carve-out
The audit's structural pass (same doctrine as the V10 god-file pass):

- window-deriver.service.ts 850→553: the prompts and the LLM-call
  cluster (base pass + truncation retry, completion-pass union,
  volume-neutral salience grading) move to src/admin/deriver-client.ts
  behind one callDeriver(deps, args) — the exact generator-client/
  verifier seam. The service keeps a thin adapter and re-exports the
  prompt symbols, so every existing import stays put. It was the next
  file over the 800 raw-line ceiling, passing only via skipComments —
  the failure shape synthesize.service.ts had at 916.
- synthesize.service.ts 759→601: the verdict/exit matrix
  (finalizeVerdict, unverifiedReturn, coverageAbstention) moves to
  src/synthesize/verdict.ts as deps-injected functions; adapters keep
  the call sites and spec bindings unchanged. File-budget headroom
  before the V11 features land.
- flag-budget gate: the RETRIEVAL_ prefix is now a DOCUMENTED,
  BOUNDED carve-out (owner decision 2026-08-09) — the gate's comment
  no longer contradicts its regex, and a second golden
  (test/golden/retrieval-profile-keys.golden.json, 19 keys) fails on
  any new RETRIEVAL_ key: profile-field status is a reviewed decision,
  not a free pass around the budget.
- ordering-frame spec: require() → import (the one lint:ci error
  sitting latent on the un-run #270 CI).

Behavior byte-identical; 238 suites / 2021 tests green, build green.
