+++ test/golden/retrieval-profile-keys.golden.json
@@ -1,0 +1,21 @@
+[
+  "RETRIEVAL_ABSTENTION_CALIBRATION",
+  "RETRIEVAL_ABSTENTION_MIN_EVIDENCE",
+  "RETRIEVAL_ABSTENTION_MIN_SCORE",
+  "RETRIEVAL_COVERAGE_SCAN_MODE",
+  "RETRIEVAL_DATE_ANCHORING",
+  "RETRIEVAL_DERIVED_VERSION",
+  "RETRIEVAL_ENTITY_EXPANSION",
+  "RETRIEVAL_GENRE",
+  "RETRIEVAL_INSIGHT_EVIDENCE",
+  "RETRIEVAL_ORDERING_FRAME",
+  "RETRIEVAL_PROFILE_OVERRIDES",
+  "RETRIEVAL_SALIENCE_SCORING",
+  "RETRIEVAL_SCAN_HNSW_EF",
+  "RETRIEVAL_SCAN_HNSW_OVERFETCH",
+  "RETRIEVAL_TEMPORAL_MODE",
+  "RETRIEVAL_TIMELINE_EVIDENCE",
+  "RETRIEVAL_UPDATE_STORY",
+  "RETRIEVAL_VERBATIM_EVIDENCE",
+  "RETRIEVAL_VERIFIER_TOPIC_COVERAGE"
+]
